Simplify and fix nextup method
It would return the right info for a talk that already started, but
return less and different kind of data (negative number of seconds until
talk instead of duration) for talks that did not start yet.

This fixes this by just keeping the dictionary for the talk and having a
single place to return values from this dict, in both cases. This
unifies the results and also slightly simplifies the code.